[mojo-stdlib] Relax the 'atol' and 'rtol' typse of isclose/approx_equal.
This changes the atol/rtol arguments to "close by" checks to take
`Float64` instead of being constrained to having to be the same as
the values being compared.  This constraint adds a bunch of casts
and makes the API more difficult to work with.  All float types
convert to Float64, so this is more flexible.

MATH_ORIG_REV_ID: 5604d2c073e916f1fc38755c451bbe5db5971449

+++ mojo/stdlib/test/math/test_math.mojo
@@ -240,7 +240,7 @@ def test_isqrt():
     assert_almost_equal(s2_f64[3], 0.89442)
 
 
-def _test_frexp_impl[type: DType](*, atol: Float32, rtol: Float32):
+def _test_frexp_impl[type: DType](*, atol: Float64, rtol: Float64):
     var res0 = frexp(Scalar[type](123.45))
     assert_almost_equal(
         res0[0].cast[DType.float32](), 0.964453, atol=atol, rtol=rtol
@@ -280,7 +280,7 @@ def _test_frexp_impl[type: DType](*, atol: Float32, rtol: Float32):
     )
 
 
-def _test_log_impl[type: DType](*, atol: Float32, rtol: Float32):
+def _test_log_impl[type: DType](*, atol: Float64, rtol: Float64):
     var res0 = log(Scalar[type](123.45))
     assert_almost_equal(
         res0.cast[DType.float32](), 4.8158, atol=atol, rtol=rtol
@@ -306,7 +306,7 @@ def _test_log_impl[type: DType](*, atol: Float32, rtol: Float32):
     assert_equal(res4, SIMD[DType.bool, 4](True, False, True, True))
 
 
-def _test_log2_impl[type: DType](*, atol: Float32, rtol: Float32):
+def _test_log2_impl[type: DType](*, atol: Float64, rtol: Float64):
     var res0 = log2(Scalar[type](123.45))
     assert_almost_equal(
         res0.cast[DType.float32](), 6.9477, atol=atol, rtol=rtol
